refactor(player-portal): decompose Inventory tab into focused files
Split the 1 026-line Inventory.tsx monolith into five focused modules
under a new `inventory/` subdirectory, keeping the original path as a
thin re-export so no consumer imports change.

- inventory-categories.ts  — InventoryCategory type, CATEGORY_ORDER/LABEL, categoryOf, groupByCategory
- inventory-shop.ts        — SellContext / InvestContext interfaces + spendCoins / grantCoins / applyCoinChanges
- InventoryItemRow.tsx     — ItemRow, ContainerChildRow, GridTile, ItemDescription, badges, buttons
- InventoryControls.tsx    — CoinStrip, ViewToggle, ShopViewToggle, ShopGearMenu
- Inventory.tsx            — tab shell (state, handlers, CategorizedInventory layout)
- index.ts                 — re-exports Inventory for the thin wrapper

Adds Vitest coverage for the pure helpers in inventory-categories.ts
(categoryOf and groupByCategory). All 413 existing tests still pass.

Future: the Party Stash section will land as a sibling component
(inventory/PartyStash.tsx) — noted in the tab shell comment.
