+++ apps/player-portal/src/components/tabs/inventory/index.ts
@@ -1,0 +1,1 @@
+export { Inventory } from './Inventory';
